Task 1: Wine Quality Prediction
🍷 Wine Quality Prediction 📈

Welcome to the Wine Quality Prediction project! This project is focused on using machine learning to predict the quality of wines based on various features.

🔍 About the Project:
-----------------------
In this project, I have explored a dataset containing a collection of red and white wine samples, along with several chemical properties and quality ratings.